# swagger/docs/swagger.yaml
definitions:
...
  dao.UserInfo:
    properties:
      avatar:
        description: 头像
        type: string
      birthday:
        description: 出生日期
        type: string
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      roles:
        description: 角色，1：学校，2：专家，3：教育局，4：教师，5：超级管理员
        items:
          type: integer
        type: array
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      subject_code:
        description: 科目code
        type: string
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户名称
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
      year:
        description: 年份
        type: integer
    type: object
...
  req.SetExpertAuthReq:
    properties:
      two_ids:
        description: 二级指标ids
        items:
          type: integer
        type: array
      user_id:
        description: 用户id
        type: string
    required:
    - two_ids
    - user_id
    type: object
  req.SetRoleToUserReq:
    properties:
      roles:
        description: 角色列表，1：学校，2：专家，3：教育局，4：老师，5：管理员
        items:
          type: integer
        type: array
      user_ids:
        description: 用户id列表
        items:
          type: string
        type: array
    required:
    - roles
    - user_ids
    type: object
...
  resp.ExpertAuthList:
    properties:
      auth_day:
        description: 授权日期，格式2006-01-02
        type: string
      expert_auth_indicator_list:
        description: 专家授权指标
        items:
          $ref: '#/definitions/resp.OneIndicatorOnlyName'
        type: array
      export_auth:
        description: 专家是否授权 0：未授权 1：已授权
        type: integer
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户名称
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
    type: object
...
  resp.GetExpertAuthListByWhereResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.ExpertAuthList'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetUserInfoResp:
    properties:
      avatar:
        description: 头像
        type: string
      birthday:
        description: 出生日期
        type: string
      create_time:
        description: 创建时间
        type: string
      declare_type:
        description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        type: integer
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      roles:
        description: 角色，1：学校，2：专家，3：教育局，4：教师，5：超级管理员
        items:
          type: integer
        type: array
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      subject_code:
        description: 科目code
        type: string
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户名称
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
      year:
        description: 年份
        type: integer
    type: object
  resp.GetUserListByWhereResp:
    properties:
      list:
        items:
          $ref: '#/definitions/dao.UserInfo'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetUsersByNameResp:
    properties:
      total:
        type: integer
      user_list:
        items:
          $ref: '#/definitions/resp.UserInfo'
        type: array
    type: object
...
  resp.OneIndicatorOnlyName:
    properties:
      one_indicator_id:
        type: integer
      one_indicator_name:
        type: string
      two_list:
        description: 二级指标列表
        items:
          $ref: '#/definitions/resp.TwoIndicatorOnlyName'
        type: array
    type: object
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error_stack: {}
      msg:
        type: string
    type: object
...
  resp.TwoIndicatorOnlyName:
    properties:
      two_indicator_id:
        description: 二级指标id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
    type: object
...
  resp.UserInfo:
    properties:
      identity_card:
        description: 身份证号
        type: string
      phone:
        description: 手机号
        type: string
      roles:
        description: 角色列表，1：学校，2：专家，3：教育局，4：老师，5：管理员
        items:
          type: integer
        type: array
      school_id:
        description: 学校id
        type: string
      school_name:
        description: 学校名称
        type: string
      user_id:
        description: 用户id
        type: string
      user_name:
        description: 用户名称
        type: string
      user_sex:
        description: 1：男，2：女
        type: integer
    type: object
...
host: localhost:7850
info:
  contact:
    email: support@swagger_controller.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a teachers-awards server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger teachers-awards API
  version: "1.0"
paths:
...
  /v1/user/cancel/expert/auth:
    delete:
      parameters:
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 取消专家授权
      tags:
      - User
  /v1/user/get/expert/auth/list:
    get:
      parameters:
      - description: 用户名称
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 二级指标ids，, 隔开
        in: query
        name: two_indicator_ids
        type: string
      - description: 专家是否授权 0：全部 1：未授权 2：已授权
        in: query
        name: export_auth
        type: integer
      - description: 授权日期，格式2006-01-02
        in: query
        name: auth_day
        type: string
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetExpertAuthListByWhereResp'
              type: object
      summary: 获取专家角色列表
      tags:
      - User
  /v1/user/get/info:
    get:
      parameters:
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetUserInfoResp'
              type: object
      summary: 获取用户信息
      tags:
      - User
  /v1/user/get/role/list:
    get:
      parameters:
      - description: 用户名称
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 手机号
        in: query
        name: phone
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 0：全部，1：学校，2：专家，3：教育局，5：超级管理员
        in: query
        name: role
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetUserListByWhereResp'
              type: object
      summary: 获取已设置角色的用户列表
      tags:
      - User
  /v1/user/list/by/name:
    get:
      parameters:
      - description: 用户名称
        in: query
        name: user_name
        required: true
        type: string
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      - description: 每页大小
        in: query
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetUsersByNameResp'
              type: object
      summary: 通过用户名称模糊匹配获取用户列表
      tags:
      - User
  /v1/user/save/info:
    post:
      parameters:
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: string
      - description: 用户名称
        in: formData
        name: user_name
        required: true
        type: string
      - description: 1：男，2：女
        in: formData
        name: user_sex
        required: true
        type: integer
      - description: 出生日期
        in: formData
        name: birthday
        required: true
        type: string
      - description: 身份证号
        in: formData
        name: identity_card
        required: true
        type: string
      - description: 手机号
        in: formData
        name: phone
        required: true
        type: string
      - description: 头像
        in: formData
        name: avatar
        required: true
        type: string
      - description: 科目code
        in: formData
        name: subject_code
        required: true
        type: string
      - description: 学校id
        in: formData
        name: school_id
        required: true
        type: string
      - description: 学校名称
        in: formData
        name: school_name
        required: true
        type: string
      - description: 1:幼教、2:小学、3:初中、4:高中、5:职校、6:教研
        in: formData
        name: declare_type
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 保存当前用户信息
      tags:
      - User
  /v1/user/set/expert/auth:
    post:
      parameters:
      - description: body请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.SetExpertAuthReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 给专家授权指标项
      tags:
      - User
  /v1/user/set/role:
    post:
      parameters:
      - description: body请求体
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/req.SetRoleToUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 设置用户角色
      tags:
      - User
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"